Fix rendering of fnumber
FNumber is represented as a imagemeta.rat[uint32], which contains a numerator and a denominator. If used directly in a template this results in fractional apertures (e.g. 5.6) in their rational form (e.g. 28/5) rather than as a float with one sig fig. 

This commit updates the EXIF rendering logic, casting the rational to a float64, and then rendering it with at most 1 sig fig